Endpoint -> registrar pontuação do Usuário

+++ src/main/java/br/com/fiap/reciclamais/gateway/repository/data/PontuacaoDocument.java
@@ -16,6 +16,6 @@ public class PontuacaoDocument {
 
     private Double reciclagem;
     private Integer total;
-    private Float percentual;
+    private Double percentual;
     private List<HistoricoDocument> historico;
 }
